requirements.txt
openai==1.35.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
numpy==2.0.0
scikit-learn==1.5.0